+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Fixed
+- Increased maximum password length from 19 to 128 characters
+
 ## [0.4.0] - 2026-05-28
 
 ### Fixed